requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blessed==1.20.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.30
llama_cpp_python==0.3.8
load-dotenv==0.1.0
marionette-driver==3.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mozdevice==4.2.0
mozfile==3.0.0
mozinfo==1.2.3
mozlog==8.0.0
mozprocess==1.4.0
mozprofile==3.0.0
mozrunner==8.3.2
mozterm==1.0.0
mozversion==2.4.0
mpmath==1.3.0
narwhals==1.34.1
neo4j==5.28.1
networkx==3.2.1
numpy==2.0.2
openai==1.70.0
opencv-python==4.11.0.86
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.3
selenium==4.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
starlette==0.46.1
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
tinydb==4.8.2
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
zstandard==0.23.0
security==1.3.1
